FAT-1740 Удалены классы CursorWrapper'ы
