+++ app/src/main/java/com/andersen/dogsapp/dogs/data/DataRepository.java
@@ -2,8 +2,8 @@ package com.andersen.dogsapp.dogs.data;
 
 import java.util.List;
 
-import com.andersen.dogsapp.dogs.data.get_entities_interfaces.IDogsDataSource;
-import com.andersen.dogsapp.dogs.data.get_entities_interfaces.IOwnersDataSource;
+import com.andersen.dogsapp.dogs.data.interfaces.IDogsDataSource;
+import com.andersen.dogsapp.dogs.data.interfaces.IOwnersDataSource;
 import com.andersen.dogsapp.dogs.data.entities.Dog;
 import com.andersen.dogsapp.dogs.data.entities.Owner;
