+++ app/src/main/java/com/andersen/dogsapp/dogs/data/json/datasources/JsonDogsDataSource.java
@@ -1,8 +1,8 @@
-package com.andersen.dogsapp.dogs.data.json_data_sources;
+package com.andersen.dogsapp.dogs.data.json.datasources;
 
 import android.content.Context;
 
-import com.andersen.dogsapp.dogs.data.get_entities_interfaces.IDogsDataSource;
+import com.andersen.dogsapp.dogs.data.interfaces.IDogsDataSource;
 import com.andersen.dogsapp.dogs.data.entities.Dog;
 import com.andersen.dogsapp.dogs.data.entities.Owner;
 import com.andersen.dogsapp.dogs.data.json.DogsData;
